+++ src/main/java/org/pethotel/HeavenForPets/mappers/impl/FoodMapImpl.java
@@ -33,7 +33,7 @@ public class FoodMapImpl implements FoodMap {
         food.setFoodType(foodEntity.getFoodType());
         food.setTaste(foodEntity.getTaste());
         food.setAmount(foodEntity.getAmount());
-        food.setId((int) foodEntity.getId());
+        food.setId(foodEntity.getId());
         return food;
     }
 }
